Finish code to correct test

+++ Structural/Decorator/CoffeeShop/src/AbstractBeverage.php
@@ -25,7 +25,7 @@ abstract class AbstractBeverage
     /**
      * Caculate beverage cost.
      *
-     * @return int
+     * @return float
      */
-    abstract public function cost(): int;
+    abstract public function cost(): float;
 }
